requirements.txt
fastapi[standard]
alembic
sqlalchemy
psycopg2-binary
python-dotenv
sqlmodel